# lib/api-zod/src/generated/types/taskInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Team Task Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { TaskInputPriority } from './taskInputPriority';
import type { TaskInputStatus } from './taskInputStatus';

export interface TaskInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  status?: TaskInputStatus;
  priority?: TaskInputPriority;
  dueDate?: string;
  projectId: number;
  assignedToId?: number;
}